+++ r4/src/test/java/io/github/jy95/fds/r4/translators/DayOfWeekTest.java
@@ -1,49 +1,31 @@
 package io.github.jy95.fds.r4.translators;
 
-import io.github.jy95.fds.r4.DosageAPIR4;
-import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.common.types.DosageAPI;
+import io.github.jy95.fds.r4.DosageAPIR4;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractDayOfWeekTest;
 import org.hl7.fhir.r4.model.Dosage;
 import org.hl7.fhir.r4.model.Timing;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
-import java.util.concurrent.ExecutionException;
 
-import static org.junit.jupiter.api.Assertions.*;
+public class DayOfWeekTest extends AbstractDayOfWeekTest<FDSConfigR4, Dosage> {
 
-public class DayOfWeekTest extends AbstractFhirTest {
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoDayOfWeek(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DAY_OF_WEEK);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testSingleDayOfWeek(Locale locale) throws ExecutionException, InterruptedException {
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DAY_OF_WEEK);
+    @Override
+    protected Dosage generateSingleDayOfWeek() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent repeatComponent = new Timing.TimingRepeatComponent();
         repeatComponent.addDayOfWeek(Timing.DayOfWeek.FRI);
         timing.setRepeat(repeatComponent);
         dosage.setTiming(timing);
-
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedSingleDayText(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testMultipleDayOfWeek(Locale locale) throws ExecutionException, InterruptedException {
-        DosageAPIR4 dosageUtils = DayOfWeekTest.getDosageAPI(locale, DisplayOrder.DAY_OF_WEEK);
+    @Override
+    protected Dosage generateMultipleDayOfWeek() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent repeatComponent = new Timing.TimingRepeatComponent();
@@ -51,35 +33,24 @@ public class DayOfWeekTest extends AbstractFhirTest {
         repeatComponent.addDayOfWeek(Timing.DayOfWeek.FRI);
         timing.setRepeat(repeatComponent);
         dosage.setTiming(timing);
-
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expectedResult = getExpectedMultipleDaysText(locale);
-        assertEquals(expectedResult, result);
+        return dosage;
     }
 
-    // For the parametrized test of single form
-    private static String getExpectedSingleDayText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "on Friday";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "le vendredi";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "am Freitag";
-        } else {
-            return "op vrijdag";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
     }
 
-    // For the parametrized test of multiple form
-    private String getExpectedMultipleDaysText(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "on Monday and Friday";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "les lundi et vendredi";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "am Montag und Freitag";
-        } else {
-            return "op maandag en vrijdag";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
+    }
+
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 }
